@@ -21,7 +21,6 @@ classifiers = [
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3 :: Only",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
@@ -29,34 +28,52 @@ classifiers = [
     "Typing :: Typed",
 ]
 
-requires-python = ">=3.8,<4.0"
-dependencies = ["ipykernel>=6", "jupyter-client>=7", "msgspec>=0.18"]
+requires-python = ">=3.9,<4.0"
+dependencies = ["jupyter-client>=7", "msgspec>=0.18"]
 
 [project.urls]
 Documentation = "https://github.com/renan-r-santos/pixi-kernel"
 Homepage = "https://github.com/renan-r-santos/pixi-kernel"
 Repository = "https://github.com/renan-r-santos/pixi-kernel"
 
+[tool.uv]
+dev-dependencies = [
+    "jupyter-kernel-test>=0.7,<0.8",
+    "mypy>=1,<2",
+    "pytest>=8,<9",
+    "pytest-cov>=5,<6",
+    "ruff>=0.6,<0.7",
+    "tox-uv>=1,<2",
+]
+
 [project.entry-points."jupyter_client.kernel_provisioners"]
-pixi-provisioner = "pixi_kernel.provisioner:PixiProvisioner"
+pixi-kernel-provisioner = "pixi_kernel.provisioner:PixiKernelProvisioner"
 
 [tool.ruff]
 src = ["src"]
 line-length = 99
+target-version = "py39"
 
 [tool.ruff.lint]
 extend-select = [
     "I", # isort
     "N", # pep8-naming
     "RUF", # ruff
+    "ASYNC", # flake8-async
     "B", # flake8-bugbear
     "N", # flake8-broken-line
     "C4", # flake8-comprehensions
+    "D", # pydocstyle
     "PIE", # flake8-pie
     "PT", # flake8-pytest-style
     "PTH", # flake8-use-pathlib
     "ERA", # flake8-eradicate
 ]
+
+# D1xx: Missing docstrings
+#`one-blank-line-before-class` (D203) and `no-blank-line-before-class` (D211) are incompatible.
+#`multi-line-summary-first-line` (D212) and `multi-line-summary-second-line` (D213) are incompatible.
+extend-ignore = ["D1", "D203", "D213"]
 
 [tool.mypy]
 strict = true
@@ -64,10 +81,13 @@ strict = true
 [tool.pytest.ini_options]
 addopts = ["--strict-config", "--strict-markers"]
 xfail_strict = true
-filterwarnings = [
-    "error",
-    "ignore:Jupyter is migrating its paths to use standard platformdirs",
-]
+
+[tool.coverage.run]
+branch = true
+source = ["src"]
+
+[tool.coverage.report]
+skip_empty = true
 
 [tool.hatch.build.targets.wheel.shared-data]
 "kernels" = "share/jupyter/kernels"
