@@ -1,108 +1,120 @@
 from __future__ import annotations
 
-import logging
 import shutil
 import subprocess
 from pathlib import Path
-from typing import List, Optional, Tuple
+from typing import Optional
 
 import msgspec
 
-from .errors import (
-    PIXI_KERNEL_NOT_FOUND,
-    PIXI_MANIFEST_NOT_FOUND,
-    PIXI_NOT_FOUND,
-    PIXI_VERSION_ERROR,
-    PIXI_VERSION_NOT_SUPPORTED,
-)
+from .errors import PIXI_KERNEL_NOT_FOUND, PIXI_NOT_FOUND, PIXI_OUTDATED, PIXI_VERSION_ERROR
 
-logger = logging.getLogger(__name__)
-MINIMUM_PIXI_VERSION = "0.21.0"
-
-
-class PixiDiscoveryError(Exception):
-    def __init__(self, message: str):
-        self.message = message
+MINIMUM_PIXI_VERSION = "0.30.0"
 
 
 class PixiInfo(msgspec.Struct, frozen=True, kw_only=True):
-    environments_info: List[EnvironmentInfo]
+    environments_info: list[EnvironmentInfo]
     project_info: Optional[ProjectInfo]
 
 
 class EnvironmentInfo(msgspec.Struct, frozen=True, kw_only=True):
     name: str
-    dependencies: List[str]
-    pypi_dependencies: List[str]
+    dependencies: list[str]
+    pypi_dependencies: list[str]
+    prefix: str
 
 
 class ProjectInfo(msgspec.Struct, frozen=True, kw_only=True):
     manifest_path: str
 
 
-def find_pixi_version(kernel_display_name: str) -> Tuple[int, int, int]:
+def ensure_readiness(*, cwd: Path, required_package: str, kernel_name: str) -> str:
+    """Ensure the Pixi environment is ready to run the kernel.
+
+    This function checks the following:
+        - Ensure Pixi is installed and in PATH
+        - Ensure the installed Pixi version is supported
+        - Ensure there is a Pixi project in the current working directory or any of its parents
+        - Ensure the Pixi project is valid and has a default environment
+        - Ensure the required kernel package is a project dependency
+
+    If any of the checks fail, a RuntimeError is raised and JupterLab will display a dialog with
+    the error message.
+
+    Returns the path to the Pixi environment prefix.
+    """
     # Ensure Pixi is in PATH
     if shutil.which("pixi") is None:
-        raise PixiDiscoveryError(PIXI_NOT_FOUND.format(kernel_display_name=kernel_display_name))
+        raise RuntimeError(PIXI_NOT_FOUND.format(kernel_name=kernel_name))
 
     # Ensure a supported Pixi version is installed
     result = subprocess.run(["pixi", "--version"], capture_output=True, text=True)
     if result.returncode != 0 or not result.stdout.startswith("pixi "):
-        raise PixiDiscoveryError(
-            PIXI_VERSION_ERROR.format(kernel_display_name=kernel_display_name)
-        )
+        raise RuntimeError(PIXI_VERSION_ERROR.format(kernel_name=kernel_name))
 
+    # Parse Pixi version and check it against the minimum required version
     pixi_version = result.stdout[len("pixi ") :].strip()
+
     major, minor, patch = map(int, pixi_version.split("."))
     required_major, required_minor, required_patch = map(int, MINIMUM_PIXI_VERSION.split("."))
+
     if (major, minor, patch) < (required_major, required_minor, required_patch):
-        raise PixiDiscoveryError(
-            PIXI_VERSION_NOT_SUPPORTED.format(
-                kernel_display_name=kernel_display_name,
-                minimum_version=MINIMUM_PIXI_VERSION,
+        raise RuntimeError(
+            PIXI_OUTDATED.format(kernel_name=kernel_name, minimum_version=MINIMUM_PIXI_VERSION)
+        )
+
+    # Ensure there is a Pixi project in the current working directory or any of its parents
+    result = subprocess.run(
+        ["pixi", "info", "--json"],
+        cwd=str(cwd.absolute()),
+        capture_output=True,
+        text=True,
+    )
+    if result.returncode != 0:
+        raise RuntimeError(f"Failed to run 'pixi info': {result.stderr}")
+
+    try:
+        pixi_info = msgspec.json.decode(result.stdout, type=PixiInfo)
+    except msgspec.MsgspecError as exception:
+        raise RuntimeError(
+            f"Failed to parse 'pixi info' output: {result.stdout}\n{exception}"
+        ) from exception
+
+    if pixi_info.project_info is None:
+        # Attempt to get a good error message by running `pixi project version get`. Maybe there's
+        # a typo in the toml file (parsing error) or there is no project at all.
+        result = subprocess.run(
+            ["pixi", "project", "version", "get"],
+            cwd=str(cwd.absolute()),
+            capture_output=True,
+            text=True,
+        )
+        raise RuntimeError(result.stderr)
+
+    # Find the default environment and check if the required kernel package is a dependency
+    for env in pixi_info.environments_info:
+        if env.name == "default":
+            default_environment = env
+            break
+    else:
+        raise RuntimeError("Default Pixi environment not found.")
+
+    dependencies = default_environment.dependencies + default_environment.pypi_dependencies
+    if required_package not in dependencies:
+        raise RuntimeError(
+            PIXI_KERNEL_NOT_FOUND.format(
+                kernel_name=kernel_name, required_package=required_package
             )
         )
-    logger.info(f"found Pixi {pixi_version}")
-    return (major, minor, patch)
 
+    # Make sure the environment can be solved and is up-to-date
+    result = subprocess.run(
+        ["pixi", "install"],
+        cwd=str(cwd.absolute()),
+        capture_output=True,
+        text=True,
+    )
+    if result.returncode != 0:
+        raise RuntimeError(f"Failed to run 'pixi install': {result.stderr}")
 
-def find_project_manifest(*, cwd: Path, package_name: str, kernel_display_name: str) -> Path:
-    # Find project's manifest file
-    candidate_dirs = [cwd, *cwd.parents]
-    for dir in candidate_dirs:
-        for project_filename in ["pixi.toml", "pyproject.toml"]:
-            result = subprocess.run(
-                ["pixi", "info", f"--manifest-path={project_filename}", "--json"],
-                cwd=dir,
-                capture_output=True,
-                text=True,
-            )
-            if result.returncode != 0:
-                logger.error(f"failed to run 'pixi info' for directory {dir}: {result.stderr}")
-                continue
-
-            try:
-                pixi_info = msgspec.json.decode(result.stdout, type=PixiInfo)
-            except msgspec.ValidationError as exception:
-                logger.error(f"failed to parse pixi info {result.stdout}: {exception}")
-                continue
-
-            if len(pixi_info.environments_info) == 0:
-                logger.warning(f"found empty project at {dir}: {pixi_info}")
-                continue
-
-            for env in pixi_info.environments_info:
-                if env.name == "default" and pixi_info.project_info is not None:
-                    if package_name not in env.dependencies + env.pypi_dependencies:
-                        logger.error(
-                            f"package {package_name} not found in project dependencies {pixi_info}"
-                        )
-                        raise PixiDiscoveryError(
-                            PIXI_KERNEL_NOT_FOUND.format(
-                                kernel_display_name=kernel_display_name,
-                                package_name=package_name,
-                            )
-                        )
-                    return dir / project_filename
-
-    raise PixiDiscoveryError(PIXI_MANIFEST_NOT_FOUND.format(cwd=cwd))
+    return default_environment.prefix
